Update defaults for DBaaS
For better monitoring and troubleshooting we change some logging parameters.
By changing to appdynamics we want to gather logs locally, as we don't
know what to do with the PostgreSQL logs yet.

Tracking of function timing and statements has been requested multiple
times already. It seems better to enable it by default.